REFACTOR: Extracted MusicSearchModule from monolithic SearchModule
